+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
 
 - Add alt text tooltips to attached audio and video.
+- Display custom emojis in profile fields.
 
 ### Changed